Fix DumpChunks in snd_mem.c for 64 bit

+++ source/snd_mem.c
@@ -236,7 +236,7 @@ void DumpChunks(void)
 #endif // PSP_VFPU
 		data_p += 4;
 		iff_chunk_len = GetLittleLong();
-		Con_Printf ("0x%x : %s (%d)\n", (int)(data_p - 4), str, iff_chunk_len);
+		Con_Printf ("0x%x : %s (%d)\n", (intptr_t)(data_p - 4), str, iff_chunk_len);
 		data_p += (iff_chunk_len + 1) & ~1;
 	} while (data_p < iff_end);
 }
@@ -339,4 +339,3 @@ wavinfo_t GetWavinfo (char *name, byte *wav, int wavlength)
 
 	return info;
 }
-
